Moved loading functions to DNNRunner
