+++ DNNTools/functions_dnn.py
@@ -31,31 +31,6 @@ def dict_to_str(dict_):
     return result.strip('_').replace('.', 'p')
 
 
-
-
-
-def load_data(inputfolder, fraction, postfix):
-
-    print(green('--> Loading data...'))
-
-
-    # os.path.join(self.inputpath_preproc, classtag, 'sample_weights_%s_train.npy' % (float_to_str(percentage)) )
-    input_train = np.load(os.path.join(inputfolder,          'input_%s_train%s.npy' % (float_to_str(fraction), postfix)))
-    input_test = np.load(os.path.join(inputfolder,           'input_%s_test%s.npy' % (float_to_str(fraction), postfix)))
-    input_val = np.load(os.path.join(inputfolder,            'input_%s_val%s.npy' % (float_to_str(fraction), postfix)))
-    labels_train = np.load(os.path.join(inputfolder,         'labels_%s_train%s.npy' % (float_to_str(fraction), postfix)))
-    labels_test = np.load(os.path.join(inputfolder,          'labels_%s_test%s.npy' % (float_to_str(fraction), postfix)))
-    labels_val = np.load(os.path.join(inputfolder,           'labels_%s_val%s.npy' % (float_to_str(fraction), postfix)))
-    sample_weights_train = np.load(os.path.join(inputfolder, 'sample_weights_%s_train%s.npy' % (float_to_str(fraction), postfix)))
-    eventweights_train = np.load(os.path.join(inputfolder,   'eventweights_%s_train%s.npy' % (float_to_str(fraction), postfix)))
-    sample_weights_test = np.load(os.path.join(inputfolder,  'sample_weights_%s_test%s.npy' % (float_to_str(fraction), postfix)))
-    eventweights_test = np.load(os.path.join(inputfolder,    'eventweights_%s_test%s.npy' % (float_to_str(fraction), postfix)))
-    sample_weights_val = np.load(os.path.join(inputfolder,   'sample_weights_%s_val%s.npy' % (float_to_str(fraction), postfix)))
-    eventweights_val = np.load(os.path.join(inputfolder,     'eventweights_%s_val%s.npy' % (float_to_str(fraction), postfix)))
-
-    return input_train, input_test, input_val, labels_train, labels_test, labels_val, sample_weights_train, sample_weights_test, sample_weights_val, eventweights_train, eventweights_test, eventweights_val
-
-
 def parameters_to_tag(parameters):
 
     tag = ''
